reuse validate_interval in results models
	modified:   allocation/__init__.py
	modified:   allocation/models/core.py
	modified:   allocation/models/results.py
	modified:   allocation/tests.py
